- ansible >= 2.2.0 - shortcut box added added $@ for ansible-playbook arguments - shortcut box-roles added argument: force|forced - update elnebuloso.server-base@2.4.0 - new role geerlingguy.ntp - new role geerlingguy.nodejs - new role ocha.yarn - removed enabling roles, using tags

+++ provision/box-roles.sh
@@ -2,6 +2,10 @@
 
 ################################################################################################
 
-ansible-galaxy install -f -r /vagrant/ansible/.box-roles.yml
+if [ "$1" = "force" ] || [ "$1" = "forced" ]; then
+    ansible-galaxy install -f -r /vagrant/ansible/.box-roles.yml
+else
+    ansible-galaxy install -r /vagrant/ansible/.box-roles.yml
+fi
 
 ################################################################################################
